+++ src/buffer_cache/large_buf.cc
@@ -588,7 +588,7 @@ uint16_t large_buf_t::segment_size(int ix) {
     }
 }
 
-buf_t *large_buf_t::get_segment_buf(int ix, uint16_t *seg_size, uint16_t *seg_offset) {
+buf_t *large_buf_t::get_segment_buf(int64_t ix, uint16_t *seg_size, uint16_t *seg_offset) {
 
     int64_t pos = floor_aligned(root_ref.offset, num_leaf_bytes()) + ix * num_leaf_bytes();
 
@@ -609,7 +609,8 @@ buf_t *large_buf_t::get_segment_buf(int ix, uint16_t *seg_size, uint16_t *seg_of
     return tr->buf;
 }
 
-const byte *large_buf_t::get_segment(int ix, uint16_t *seg_size) {
+// TODO check that people use int64_t for ix.
+const byte *large_buf_t::get_segment(int64_t ix, uint16_t *seg_size) {
     assert(state == loaded);
     assert(ix >= 0 && ix < get_num_segments());
 
@@ -623,7 +624,8 @@ const byte *large_buf_t::get_segment(int ix, uint16_t *seg_size) {
     return leaf->buf + seg_offset;
 }
 
-byte *large_buf_t::get_segment_write(int ix, uint16_t *seg_size) {
+// TODO check that people use int64_t for ix.
+byte *large_buf_t::get_segment_write(int64_t ix, uint16_t *seg_size) {
     assert(state == loaded);
     assert(ix >= 0 && ix < get_num_segments());
